layout: moved engine/console to cmd/snakep2p/tui.
