+++ cmd/snakep2p/tui/game.go
@@ -1,4 +1,4 @@
-package console
+package tui
 
 import (
 	"fmt"
